Adiciona validação dos dados

+++ src/Controllers/ProviderController.php
@@ -1,6 +1,7 @@
 <?php
     require_once(__DIR__.'/../../autoload.php');
     require_once('src/helpers/SessionValidate.php');
+    require_once('src/helpers/Validate.php');
 
     class ProviderController{
         public function index(){
@@ -12,14 +13,33 @@
         }
 
         public function store($params){
+            $data = array(
+                'Nome' => $params['nome'],
+                'CNPJ' => $params['cnpj'],
+                'Endereco' => $params['endereco'],
+                'Telefone' => $params['telefone'],
+                'Email' => $params['email']
+            );
+            $validator = array(
+                'Nome' => 'required',
+                'CNPJ' => 'required|cnpj',
+                'Endereco' => 'required',
+                'Telefone' => 'required',
+                'Email' => 'required|email',
+            );
+            $resultValidation = validate($data,$validator);
             if($_SESSION['dados_usuario']['nivelAcesso'] == 2){
-                $provider = new ProviderModel();
-                $result = $provider->create($params);
-
-                if($result){
-                    return redirect('/add/provider', 'Forncedor cadastrado com sucesso');
+                if($resultValidation['success']){
+                    $provider = new ProviderModel();
+                    $result = $provider->create($params);
+    
+                    if($result){
+                        return redirect('/provider', 'Forncedor cadastrado com sucesso');
+                    }else{
+                        return redirect('/add/provider', 'Ocorreu um erro interno. Contate os desenvolvedores.');
+                    }
                 }else{
-                    return redirect('/add/provider', 'Algo deu errado, tente novamente');
+                    return redirect('/provider', $resultValidation['message']);
                 }
             }else{
                 return redirect('/provider', 'Usuário sem permissão.');
@@ -53,29 +73,62 @@
         }
 
         public function update($params){
+            $data = array(
+                'Nome' => $params['nome'],
+                'CNPJ' => $params['cnpj'],
+                'Endereco' => $params['endereco'],
+                'Telefone' => $params['telefone'],
+                'Email' => $params['email'],
+                'ID_do_fornecedor' => $params['providerId']
+            );
+            $validator = array(
+                'Nome' => 'required',
+                'CNPJ' => 'required|cnpj',
+                'Endereco' => 'required',
+                'Telefone' => 'required',
+                'Email' => 'required|email',
+                'ID_do_fornecedor' => 'required'
+            );
+            $resultValidation = validate($data,$validator);
             if($_SESSION['dados_usuario']['nivelAcesso'] == 2){
-                $provider = new ProviderModel();
-                $result = $provider->update($params);
+                if($resultValidation['success']){
+                    $provider = new ProviderModel();
+                    $result = $provider->update($params);
 
-                if($result){
-                    return redirect('/provider', 'Fornecedor atualizado com sucesso');
+                    if($result){
+                        return redirect('/provider', 'Fornecedor atualizado com sucesso');
+                    }else{
+                        return redirect('/provider', 'Ocorreu um erro interno. Contate os desenvolvedores.');
+                    }
                 }else{
-                    return redirect('/provider', 'Algo deu errado, tente novamente');
+                    return redirect('/provider', $resultValidation['message']);
                 }
             }else{
                 return redirect('/provider', 'Usuário sem permissão.');
             }
         }
         public function delete($params){
+            $data = array(
+                'ID_do_fornecedor' => $params['id']
+            );
+            $validator = array(
+                'ID_do_fornecedor' => 'required'
+            );
+            $resultValidation = validate($data,$validator);
             if($_SESSION['dados_usuario']['nivelAcesso'] == 2){
-                $provider = new ProviderModel();
-                $result = $provider->delete($params);
+                if($resultValidation['success']){
+                    $provider = new ProviderModel();
+                    $result = $provider->delete($params);
 
-                if($result){
-                    return redirect('/provider', 'Fornecedor deletado com sucesso');
+                    if($result){
+                        return redirect('/provider', 'Fornecedor deletado com sucesso');
+                    }else{
+                        return redirect('/provider', 'Ocorreu um erro interno. Contate os desenvolvedores.');
+                    }
                 }else{
-                    return redirect('/provider', 'Algo deu errado, tente novamente');
+                    return redirect('/provider', $resultValidation['message']);
                 }
+
             }else{
                 return redirect('/provider', 'Usuário sem permissão.');
             }
